drivers/flash: Adapt to possible wide off_t type
off_t is a POSIX type used for sizes of files. Zephyr uses it for the
position of objects within flash storage.

It assumes that off_t is equivalent to 'long' when printing offset values
as well as being the same size as a pointer when converting offsets to
addresses for in-memory flash regions.

Neither of these assumptions hold when off_t is 64-bits on 32-bit targets.

When used as a pointer offset, cast values to ptrdiff_t. To print them
use the 't' length specifier.

When used as a memory address, cast values to uintptr_t, including before
casting to a pointer. To print them, cast to long and use a 'l' length
specifier.

+++ drivers/flash/flash_realtek_rts5912.c
@@ -672,7 +672,7 @@ static int flash_rts5912_erase(const struct device *dev, off_t offset, size_t le
 	for (; len > 0; len -= FLASH_ERASE_BLK_SZ) {
 		ret = flash_erase_sector(dev, offset);
 		if (ret < 0) {
-			LOG_ERR("erase @0x%08lx fail", offset);
+			LOG_ERR("erase @0x%08lx fail", (long)offset);
 		}
 		offset += FLASH_ERASE_BLK_SZ;
 	}
